[move/#18] enum class 분리
