+++ app/src/main/java/org/sopt/melon/presentation/home/type/NewSongFilter.kt
@@ -1,0 +1,9 @@
+package org.sopt.melon.presentation.home.type
+
+enum class NewSongFilter(
+    val displayName: String,
+) {
+    ALL(displayName = "전체"),
+    DOMESTIC(displayName = "국내"),
+    OVERSEAS(displayName = "해외"),
+}
